refactor: App, Column 컴포넌트 분리
