+++ src/components/Column.jsx
@@ -1,0 +1,51 @@
+import React from "react";
+import { Draggable, Droppable } from "react-beautiful-dnd";
+
+export default function Column({ columnOrder, columns }) {
+  const GRID = 8;
+
+  const getItemStyle = (isDragging, draggableStyle) => ({
+    userSelect: "none",
+    padding: GRID * 2,
+    margin: `0 0 ${GRID}px 0`,
+    background: isDragging ? "lightgreen" : "grey",
+    ...draggableStyle,
+  });
+
+  const getListStyle = (isDraggingOver) => ({
+    background: isDraggingOver ? "lightblue" : "lightgrey",
+    padding: GRID,
+    width: 250,
+  });
+
+  return (
+    <Droppable key={columnOrder} droppableId={columnOrder}>
+      {(provided, snapshot) => (
+        <div
+          {...provided.droppableProps}
+          ref={provided.innerRef}
+          style={getListStyle(snapshot.isDraggingOver)}
+        >
+          {columns[columnOrder].contents.map((item, index) => (
+            <Draggable key={item.id} draggableId={item.id} index={index}>
+              {(provided, snapshot) => (
+                <div
+                  ref={provided.innerRef}
+                  {...provided.draggableProps}
+                  {...provided.dragHandleProps}
+                  style={getItemStyle(
+                    snapshot.isDragging,
+                    provided.draggableProps.style,
+                  )}
+                >
+                  {item.content}
+                </div>
+              )}
+            </Draggable>
+          ))}
+          {provided.placeholder}
+        </div>
+      )}
+    </Droppable>
+  );
+}
